removing monaco-editor-theme package
