+++ packages/docs-app/package.json
@@ -17,13 +17,13 @@
         "verify": "run-p dist lint"
     },
     "dependencies": {
+        "@blueprintjs/colors": "workspace:^",
         "@blueprintjs/core": "workspace:^",
         "@blueprintjs/datetime": "workspace:^",
         "@blueprintjs/docs-data": "workspace:^",
         "@blueprintjs/docs-theme": "workspace:^",
         "@blueprintjs/icons": "workspace:^",
         "@blueprintjs/labs": "workspace:^",
-        "@blueprintjs/monaco-editor-theme": "workspace:^",
         "@blueprintjs/select": "workspace:^",
         "@blueprintjs/table": "workspace:^",
         "@blueprintjs/test-commons": "workspace:^",
