--- packages/monaco-editor-theme/.npmignore
@@ -1,5 +1,0 @@
-coverage/
-test/
-webpack.config.js
-karma.conf.js
-.npmrc
