--- packages/monaco-editor-theme/README.md
@@ -1,15 +1,0 @@
-<img height="204" src="https://cloud.githubusercontent.com/assets/464822/20228152/d3f36dc2-a804-11e6-80ff-51ada2d13ea7.png">
-
-# [Blueprint](http://blueprintjs.com/) Monaco Editor Theme
-
-Blueprint is a React UI toolkit for the web.
-
-This package provides a theme for [monaco-editor](https://microsoft.github.io/monaco-editor).
-
-## Installation
-
-```sh copy
-npm install --save @blueprintjs/monaco-editor-theme
-```
-
-### [Full Documentation](http://blueprintjs.com/docs) | [Source Code](https://github.com/palantir/blueprint)
